Add files via upload
Implemented 2 logistic regression machine learning algorithms. The first is used in order to assess whether or not a user would have watched a netflix film given previous filmed that they have already watched, which utilizes a test set of 19 different movies that many users watched and uses that to and whether they have watched X movie. Then, it uses this data to make future predictions on whether someone is likely to watch X movie. The second algorithm looks at user cell phone reviews, and uses the presence of negative words in order to classify as review as positive or negative. It then uses the trained data to make assessments on whether a review will be positive or negative using the words used in the review. The Netflix Algorithm is 66% accurate, and the cell-phone review algorithm is 74% accurate.